Finished features for step 3

+++ database/migrations/2021_05_13_155304_create_movements_table.php
@@ -18,7 +18,7 @@ class CreateMovementsTable extends Migration
             $table->unsignedBigInteger('user_id'); 
             $table->foreign('user_id')->references('id')->on('users'); 
             $table->enum('operation', ['credit', 'debit', 'reversal']);
-            $table->float('value', 8, 2);
+            $table->bigInteger('value');
             $table->timestamps();
         });
     }
